Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/resource-timing/crossorigin-sandwich-no-TAO.sub-expected.txt



PASS window.performance is defined
PASS window.performance.getEntriesByName is defined
PASS There should be one entry.
FAIL redirectStart == 0 in cross-origin redirect with failing Timing-Allow-Origin. assert_equals: redirectStart == 0 in cross-origin redirect with failing Timing-Allow-Origin. expected 0 but got 22
FAIL redirectEnd == 0 in cross-origin redirect with failing Timing-Allow-Origin. assert_equals: redirectEnd == 0 in cross-origin redirect with failing Timing-Allow-Origin. expected 0 but got 34
PASS fetchStart > 0 in cross-origin redirect with failing Timing-Allow-Origin.
FAIL startTime == fetchStart in cross-origin redirect with failing Timing-Allow-Origin. assert_equals: startTime == fetchStart in cross-origin redirect with failing Timing-Allow-Origin. expected 22 but got 34
